[chore] rename page file
